+++ src/View/Helper/DropdownHelper.php
@@ -64,7 +64,7 @@ class DropdownHelper extends Helper
      * Helpers
      * @var array
      */
-    public $helpers = ['Html' => ['className' => METOOLS . '.Html']];
+    public $helpers = ['Html' => ['className' => ME_TOOLS . '.Html']];
 
     /**
      * Start link. This link allows the opening of the dropdown menu